# services/tanren-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tanren",
    "description": "Tanren worker-manager HTTP API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/dispatch": {
      "post": {
        "tags": [
          "dispatch"
        ],
        "summary": "Create Dispatch",
        "description": "Accept a new dispatch request.",
        "operationId": "create_dispatch_api_v1_dispatch_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dispatch/{dispatch_id}": {
      "get": {
        "tags": [
          "dispatch"
        ],
        "summary": "Get Dispatch",
        "description": "Query dispatch status by workflow ID.",
        "operationId": "get_dispatch_api_v1_dispatch__dispatch_id__get",
        "parameters": [
          {
            "name": "dispatch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Workflow ID",
              "title": "Dispatch Id"
            },
            "description": "Workflow ID"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "dispatch"
        ],
        "summary": "Cancel Dispatch",
        "description": "Cancel a pending dispatch.",
        "operationId": "cancel_dispatch_api_v1_dispatch__dispatch_id__delete",
        "parameters": [
          {
            "name": "dispatch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Workflow ID",
              "title": "Dispatch Id"
            },
            "description": "Workflow ID"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchCancelled"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Cli": {
        "type": "string",
        "enum": [
          "opencode",
          "codex",
          "claude",
          "bash"
        ],
        "title": "Cli",
        "description": "CLI tool types from PROTOCOL.md Section 2."
      },
...
      "DispatchAccepted": {
        "properties": {
          "dispatch_id": {
            "type": "string",
            "title": "Dispatch Id",
            "description": "Auto-generated workflow identifier"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Acceptance status",
            "default": "accepted"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dispatch_id"
        ],
        "title": "DispatchAccepted",
        "description": "Thin response confirming dispatch acceptance."
      },
      "DispatchCancelled": {
        "properties": {
          "dispatch_id": {
            "type": "string",
            "title": "Dispatch Id",
            "description": "Cancelled workflow identifier"
          },
          "status": {
            "$ref": "#/components/schemas/DispatchRunStatus",
            "description": "Cancellation status",
            "default": "cancelled"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dispatch_id"
        ],
        "title": "DispatchCancelled",
        "description": "Confirmation that a dispatch was cancelled."
      },
      "DispatchDetail": {
        "properties": {
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id",
            "description": "Unique workflow identifier"
          },
          "phase": {
            "$ref": "#/components/schemas/Phase",
            "description": "Dispatch phase type"
          },
          "project": {
            "type": "string",
            "title": "Project",
            "description": "Project name"
          },
          "spec_folder": {
            "type": "string",
            "title": "Spec Folder",
            "description": "Relative path to spec folder"
          },
          "branch": {
            "type": "string",
            "title": "Branch",
            "description": "Git branch name"
          },
          "cli": {
            "$ref": "#/components/schemas/Cli",
            "description": "CLI tool used"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "Model identifier"
          },
          "timeout": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Timeout",
            "description": "Max execution time in seconds"
          },
          "environment_profile": {
            "type": "string",
            "title": "Environment Profile",
            "description": "Environment profile name"
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context",
            "description": "Extra context for the agent"
          },
          "gate_cmd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gate Cmd",
            "description": "Shell command for gate phases"
          },
          "status": {
            "$ref": "#/components/schemas/DispatchRunStatus",
            "description": "Current dispatch status"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Outcome"
              },
              {
                "type": "null"
              }
            ],
            "description": "Final outcome if completed"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO 8601 creation timestamp"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At",
            "description": "ISO 8601 start timestamp"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At",
            "description": "ISO 8601 completion timestamp"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "workflow_id",
          "phase",
          "project",
          "spec_folder",
          "branch",
          "cli",
          "timeout",
          "environment_profile",
          "status",
          "created_at"
        ],
        "title": "DispatchDetail",
        "description": "Full dispatch detail including runtime tracking."
      },
...
      "DispatchRequest": {
        "properties": {
          "project": {
            "type": "string",
            "title": "Project",
            "description": "Project name matching repo name"
          },
          "phase": {
            "$ref": "#/components/schemas/Phase",
            "description": "Dispatch phase type"
          },
          "branch": {
            "type": "string",
            "title": "Branch",
            "description": "Git branch name"
          },
          "spec_folder": {
            "type": "string",
            "title": "Spec Folder",
            "description": "Relative path to spec folder"
          },
          "cli": {
            "$ref": "#/components/schemas/Cli",
            "description": "CLI tool to use"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "Model identifier"
          },
          "timeout": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Timeout",
            "description": "Max execution time in seconds",
            "default": 1800
          },
          "environment_profile": {
            "type": "string",
            "title": "Environment Profile",
            "description": "Environment profile name",
            "default": "default"
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context",
            "description": "Extra context for the agent"
          },
          "gate_cmd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gate Cmd",
            "description": "Shell command for gate phases"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project",
          "phase",
          "branch",
          "spec_folder",
          "cli"
        ],
        "title": "DispatchRequest",
        "description": "Submit a dispatch \u2014 omits workflow_id (auto-generated server-side)."
      },
      "DispatchRunStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "DispatchRunStatus",
        "description": "Runtime status of a dispatch."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Outcome": {
        "type": "string",
        "enum": [
          "success",
          "fail",
          "blocked",
          "error",
          "timeout"
        ],
        "title": "Outcome",
        "description": "Result outcomes from PROTOCOL.md Section 3."
      },
...
      "Phase": {
        "type": "string",
        "enum": [
          "do-task",
          "audit-task",
          "run-demo",
          "audit-spec",
          "investigate",
          "gate",
          "setup",
          "cleanup"
        ],
        "title": "Phase",
        "description": "Dispatch phase types from PROTOCOL.md Section 2."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}